+++ packages/generator/__tests__/generate-token.test.ts
@@ -452,6 +452,7 @@ describe('generator', () => {
           --spacing-1\\.5: 0.375rem;
           --spacing-2\\.5: 0.625rem;
           --spacing-3\\.5: 0.875rem;
+          --spacing-4\\.5: 1.125rem;
           --sizes-0: 0rem;
           --sizes-1: 0.25rem;
           --sizes-2: 0.5rem;
@@ -486,6 +487,7 @@ describe('generator', () => {
           --sizes-1\\.5: 0.375rem;
           --sizes-2\\.5: 0.625rem;
           --sizes-3\\.5: 0.875rem;
+          --sizes-4\\.5: 1.125rem;
           --sizes-xs: 20rem;
           --sizes-sm: 24rem;
           --sizes-md: 28rem;