+++ packages/parser/__tests__/preset-patterns.test.ts
@@ -2212,6 +2212,10 @@ describe('staticCss', () => {
           --bleed-x: var(--spacing-3\\.5, \\33\\.5);
       }
 
+        .\\--bleed-x_token\\(spacing\\.4\\.5\\,_4\\.5\\) {
+          --bleed-x: var(--spacing-4\\.5, \\34\\.5);
+      }
+
         .\\--bleed-x_token\\(spacing\\.gutter\\,_gutter\\) {
           --bleed-x: var(--spacing-gutter, gutter);
       }
@@ -2346,6 +2350,10 @@ describe('staticCss', () => {
 
         .\\--bleed-x_token\\(spacing\\.-3\\.5\\,_-3\\.5\\) {
           --bleed-x: calc(var(--spacing-3\\.5) * -1, -\\33\\.5);
+      }
+
+        .\\--bleed-x_token\\(spacing\\.-4\\.5\\,_-4\\.5\\) {
+          --bleed-x: calc(var(--spacing-4\\.5) * -1, -\\34\\.5);
       }
 
         .\\--bleed-x_token\\(spacing\\.-gutter\\,_-gutter\\) {
@@ -2517,6 +2525,10 @@ describe('staticCss', () => {
 
         .\\--thickness_3\\.5 {
           --thickness: 3.5;
+      }
+
+        .\\--thickness_4\\.5 {
+          --thickness: 4.5;
       }
 
         .\\--thickness_xs {
@@ -2810,6 +2822,10 @@ describe('staticCss', () => {
           inset-inline-end: var(--spacing-3\\.5);
       }
 
+        .inset-e_4\\.5 {
+          inset-inline-end: var(--spacing-4\\.5);
+      }
+
         .inset-e_gutter {
           inset-inline-end: var(--spacing-gutter);
       }
@@ -2944,6 +2960,10 @@ describe('staticCss', () => {
 
         .inset-e_-3\\.5 {
           inset-inline-end: calc(var(--spacing-3\\.5) * -1);
+      }
+
+        .inset-e_-4\\.5 {
+          inset-inline-end: calc(var(--spacing-4\\.5) * -1);
       }
 
         .inset-e_-gutter {
@@ -3082,6 +3102,10 @@ describe('staticCss', () => {
           inset-block-start: var(--spacing-3\\.5);
       }
 
+        .inset-bs_4\\.5 {
+          inset-block-start: var(--spacing-4\\.5);
+      }
+
         .inset-bs_gutter {
           inset-block-start: var(--spacing-gutter);
       }
@@ -3216,6 +3240,10 @@ describe('staticCss', () => {
 
         .inset-bs_-3\\.5 {
           inset-block-start: calc(var(--spacing-3\\.5) * -1);
+      }
+
+        .inset-bs_-4\\.5 {
+          inset-block-start: calc(var(--spacing-4\\.5) * -1);
       }
 
         .inset-bs_-gutter {