+++ packages/studio/styled-system/tokens/index.mjs
@@ -467,6 +467,10 @@ const tokens = {
     "value": "0.875rem",
     "variable": "var(--spacing-3\\.5)"
   },
+  "spacing.4.5": {
+    "value": "1.125rem",
+    "variable": "var(--spacing-4\\.5)"
+  },
   "sizes.0": {
     "value": "0rem",
     "variable": "var(--sizes-0)"
@@ -602,6 +606,10 @@ const tokens = {
   "sizes.3.5": {
     "value": "0.875rem",
     "variable": "var(--sizes-3\\.5)"
+  },
+  "sizes.4.5": {
+    "value": "1.125rem",
+    "variable": "var(--sizes-4\\.5)"
   },
   "sizes.xs": {
     "value": "20rem",
@@ -1870,6 +1878,10 @@ const tokens = {
   "spacing.-3.5": {
     "value": "calc(var(--spacing-3\\.5) * -1)",
     "variable": "var(--spacing-3\\.5)"
+  },
+  "spacing.-4.5": {
+    "value": "calc(var(--spacing-4\\.5) * -1)",
+    "variable": "var(--spacing-4\\.5)"
   },
   "colors.colorPalette": {
     "value": "var(--colors-color-palette)",